fix: remove profile page feature flag and cleanup old profile components
