+++ src/app/[lng]/(infos)/profile/[walletAddress]/page.tsx
@@ -3,15 +3,9 @@ import { walletAddressSchema } from '@/utils/validation-schemas';
 import type { Metadata } from 'next';
 import { notFound } from 'next/navigation';
 import { Suspense } from 'react';
-import { getFeatureFlag } from 'src/app/lib/getFeatureFlag';
 import { getPerks } from 'src/app/lib/getPerks';
-import { getProfileBannerCampaigns } from 'src/app/lib/getProfileBannerCampaigns';
-import { getQuestsWithNoCampaignAttached } from 'src/app/lib/getQuestsWithNoCampaignAttached';
-import OldProfilePage from 'src/app/ui/profile/OldProfilePage';
 import { ProfilePage } from 'src/components/ProfilePage/ProfilePage';
 import { ProfilePageSkeleton } from 'src/components/ProfilePage/ProfilePageSkeleton';
-import { GlobalFeatureFlags } from 'src/const/abtests';
-import { fetchQuestOpportunitiesByRewardsIds } from 'src/utils/merkl/fetchQuestOpportunities';
 
 type Params = Promise<{ walletAddress: string }>;
 
@@ -68,39 +62,19 @@ export default async function Page({ params }: { params: Params }) {
   }
 
   const sanitizedAddress = result.data;
-  const [
-    { data: campaigns },
-    { data: questsData },
-    { data: perksResponse },
-    isPageEnabled,
-  ] = await Promise.all([
-    getProfileBannerCampaigns(),
-    getQuestsWithNoCampaignAttached(),
-    getPerks(),
-    getFeatureFlag(GlobalFeatureFlags.ProfilePage),
-  ]);
+  const { data: perksResponse } = await getPerks();
 
-  const questsExtended = await fetchQuestOpportunitiesByRewardsIds(
-    questsData.data,
-  );
-
-  if (isPageEnabled) {
-    const perks = perksResponse.data;
-    const totalPerks = perksResponse.meta.pagination?.total || 0;
-    const hasMorePerks = totalPerks > perks.length;
-
-    return (
-      <Suspense fallback={<ProfilePageSkeleton />}>
-        <ProfilePage perks={perks} hasMorePerks={hasMorePerks} />
-      </Suspense>
-    );
-  }
+  const perks = perksResponse.data;
+  const totalPerks = perksResponse.meta.pagination?.total || 0;
+  const hasMorePerks = totalPerks > perks.length;
 
   return (
-    <OldProfilePage
-      quests={questsExtended}
-      campaigns={campaigns}
-      walletAddress={sanitizedAddress}
-    />
+    <Suspense fallback={<ProfilePageSkeleton />}>
+      <ProfilePage
+        walletAddress={sanitizedAddress}
+        perks={perks}
+        hasMorePerks={hasMorePerks}
+      />
+    </Suspense>
   );
 }
