--- src/components/ProfilePage/QuestCard/VoidQuestCard.tsx
@@ -1,16 +1,0 @@
-import {
-  VoidQuestCardContainer,
-  VoidQuestCardIcon,
-} from './VoidQuestCard.style';
-
-// interface VoidQuestCardProps {
-//   connected: boolean;
-// }
-
-export const VoidQuestCard = () => {
-  return (
-    <VoidQuestCardContainer>
-      <VoidQuestCardIcon />
-    </VoidQuestCardContainer>
-  );
-};
