+++ src/components/ProfilePage/index.ts
@@ -1,3 +1,1 @@
-export * from './AddressCard/IconButton.style';
-export * from './OldProfilePage';
 export * from './ProfilePage';
